Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,18 +1,18 @@
 #include <stdio.h>
 int main (){
-    float a , b;
+    int a , b;
     char sign ;
     scanf("%f %f %c",&a,&b,&sign);
-    if(sign=='+'){
+    if(sign == '+'){
         printf("%d",a+b);
     }
-    if (sign=='-'){
+    if (sign == '-'){
         printf("%d",a-b);
     }
-    if (sign=='*' ){
+    if (sign == '*' ){
         printf("%d",a*b);
     }
-    if (sign=='/'){
+    if (sign == '/'){
         printf("%d",a/b);
     }
 }
